chore: move evaluationVersion to CommonConstants (#34200)
## Description
> Move EVALUATION_VERSION to `CommonConstanta.java`

Fixes https://github.com/appsmithorg/appsmith/issues/34171

## Automation

/ok-to-test tags="@tag.Sanity,@tag.Workflows,@tag.Git,@tag.JS"

### :mag: Cypress test results
