+++ app/server/appsmith-server/src/main/java/com/appsmith/server/constants/ce/CommonConstantsCE.java
@@ -6,4 +6,5 @@ public class CommonConstantsCE {
     public static String FOR = DELIMETER_SPACE + "for" + DELIMETER_SPACE;
     public static String IN = DELIMETER_SPACE + "in" + DELIMETER_SPACE;
     public static String DEFAULT = "default";
+    public static final Integer EVALUATION_VERSION = 2;
 }
